requirements.txt
# Core dependencies
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2

# Data collection
yfinance==0.2.33
alpaca-py==0.15.0

# Technical analysis
pandas-ta==0.3.14b0
ta-lib==0.4.28

# Database
sqlalchemy==2.0.23

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality
black==23.12.1
flake8==6.1.0
mypy==1.7.1
pylint==3.0.3

# Utilities
requests==2.31.0
pytz==2023.3.post1
python-dateutil==2.8.2

# Monitoring and notifications (optional)
streamlit==1.29.0

# Type hints
types-requests==2.31.0.10
types-pytz==2023.3.1.1
